pyproject.toml
[tool.poetry]
name = "pyvisionai"
version = "0.1.0"
description = "A Python library for extracting and describing content from documents using Vision LLMs"
authors = ["MDGrey33 <roland@abouyounes.com>"]
license = "Apache-2.0"
readme = "README.md"
repository = "https://github.com/MDGrey33/content-extractor-with-vision"
keywords = ["pdf", "docx", "pptx", "vision", "llm", "extraction"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
]
packages = [
    { include = "pyvisionai" }
]

[tool.poetry.scripts]
file-extract = "pyvisionai.cli.main:main"
describe-image = "pyvisionai.cli.describe_image:main"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
Pillow-PIL = "^0.1.dev0"
ollama = "^0.4.2"
pillow = "^11.0.0"
python-docx = "^1.1.2"
python-pptx = "^1.0.2"
black = "^24.10.0"
openai = "^1.58.1"
pdf2image = "^1.17.0"
pdfminer-six = "^20231228"
PyPDF2 = "^3.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
